src/main/java/com/CabApp/model/CabFares.java
package com.CabApp.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
@Entity
@Table(name = "Cab_Fares")
public class CabFares {
	@Id
    @Column(name = "ID")
    private int id;

    @Column(name = "FromLocation")
    private String fromLocation;

    @Column(name = "ToLocation")
    private String toLocation;

    @Column(name = "Fare")
    private int fare;

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getFromLocation() {
		return fromLocation;
	}

	public void setFromLocation(String fromLocation) {
		this.fromLocation = fromLocation;
	}

	public String getToLocation() {
		return toLocation;
	}

	public void setToLocation(String toLocation) {
		this.toLocation = toLocation;
	}

	public int getFare() {
		return fare;
	}

	public void setFare(int fare) {
		this.fare = fare;
	}
    
    
}
